+++ api/database.py
@@ -9,7 +9,6 @@ from sqlalchemy.ext.asyncio import (
 )
 from sqlalchemy.orm import declarative_base
 from typing import Annotated
-import asyncio
 import logging
 
 logger = logging.getLogger(__name__)
@@ -44,15 +43,3 @@ async def db_session() -> AsyncGenerator[AsyncSession, None]:
 
 
 Database = Annotated[AsyncSession, Depends(db_session)]
-
-
-# Migration
-async def migrate_tables() -> None:
-    logger.info("Starting to migrate")
-    async with engine.begin() as conn:
-        await conn.run_sync(Base.metadata.create_all)
-    logger.info("Done migrating")
-
-
-if __name__ == "__main__":
-    asyncio.run(migrate_tables())
