+++ api/migrate.py
@@ -1,0 +1,11 @@
+import asyncio
+from .database import Base, engine
+
+
+async def migrate_tables() -> None:
+    async with engine.begin() as conn:
+        await conn.run_sync(Base.metadata.create_all)
+
+
+if __name__ == "__main__":
+    asyncio.run(migrate_tables())
